Improve bootstrap.sh.
The script was failing if one of the directories existed already, and the current version of pathfinder is incompatible with our code (changed package name from pathfinder to addon-pathfinder) and incompatible with itself (removed addon/unload.js without fixing its own uses of it). So this checks out a known-to-work version of pathfinder instead.